PCP: fix missing pieces
Apparently the problem is that the pieces that weren't cut were too far
to the right. Despite the fact that there was cardboard to cut. I tried
updating the document properties in Inkscape by setting these to A4, but
the pieces were still not being cut. It does help to move the whole
image 7 px to the left. The solution wasn't tested on the kite shaped
faces. These might still be too far to the right. At least the x
coordinate 573 px seems to work (the diamond has an extreme of 576)